requirements.txt
freetype-py==1.0.2
gitdb2==2.0.0
GitPython==2.1.1
smmap2==2.0.1
